+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## v0.8.3
+
+### New
+
+- Added `isPrime` function which accepts a number and determines if its prime or not
+- Added `factors` function which accepts a number and finds its factors
+
+### Improved
+
+- Tweaked `range` handling and flow to work better with `factors`
+
+### Fixed
+
+- Case where passing a string number to range might produce weird results
+
 ## v0.8.2
 
 ### Fixed